requirements.txt

Django==4.2.18
django-environ==0.11.2
djangorestframework==3.16.0

pip~=25.1.1
setuptools~=65.5.0
psycopg2~=2.9.10
pexpect~=4.9.0
Unipath~=1.1
python-decouple==3.4
attrs~=25.3.0
PyYAML~=6.0.2
asgiref~=3.8.1
sqlparse~=0.5.3
typing_extensions~=4.14.0
ptyprocess~=0.7.0
referencing~=0.36.2
jsonschema~=4.24.0
uritemplate~=4.2.0
inflection~=0.5.1